marker flow restructured, my photos pg dynamic
